refactor: rename NUMERIC/NumericLiteral to INT/IntLiteral
Aligns integer literal naming with the FLOAT/FloatLiteral convention
across token types, AST nodes, formatters, and the LSP hover handler.

+++ src/spud/stage_five/stage_five_token_type.py
@@ -3,7 +3,7 @@ from enum import Enum
 
 class StageFiveTokenType(str, Enum):
     IDENTIFIER = "identifier"
-    NUMERIC = "NUMERIC"
+    INT = "INT"
     FLOAT = "FLOAT"
     NEW_LINE = "\n"
     STRING = "STRING"
